index: fix regexp->trigram behaviour when trigrams are question-marked
The code for regex alternation assumed that either x and y are both
exact or both have prefix/suffix lists.  This obviously isn't always
the case.

This particular bug is caused by the fact that emptyString() has an
exact match but no prefix/suffix lists, so when being combined we were
unioning with a nil list.  This didn't crash, it just didn't have the
expected behaviour.  We now also call addExact() before combining the
matches.

Fixes issue 17.

R=rsc
http://codereview.appspot.com/6057050